CHORE[feature/services]: adjusting general files
