+++ src/components/cardCentral/index.tsx
@@ -19,22 +19,7 @@ import {
   TopContentContainer
 } from './styles';
 import { YearIcon, KmIcon, JeepLogo, RenagadeImage } from 'assets';
-
-export interface CarData {
-  id: string;
-  brand: string;
-  name: string;
-  year: string;
-  km: string;
-  price: number;
-  specifications: string;
-  image: string;
-  color: string;
-  fuel: string;
-  fuelUrban: string;
-  fuelRoad: string;
-  dataSheet: string;
-}
+import { CarData } from 'services/interfaces';
 
 interface CardCentralProps {
   carData: CarData;
