+++ src/pages/_app.tsx
@@ -1,5 +1,6 @@
 import isPropValid from '@emotion/is-prop-valid';
 import { AppProps } from 'next/app';
+import Head from 'next/head';
 import { StyleSheetManager, ThemeProvider } from 'styled-components';
 import SimpleReactLightbox from 'simple-react-lightbox';
 
@@ -10,6 +11,9 @@ function App({ Component, pageProps }: AppProps) {
   return (
     <StyleSheetManager shouldForwardProp={isPropValid}>
       <ThemeProvider theme={theme}>
+        <Head>
+          <title>CarRent Verzel</title>
+        </Head>
         <GlobalStyles />
         <SimpleReactLightbox>
           <Component {...pageProps} />
