+++ src/services/interfaces/index.tsx
@@ -1,0 +1,15 @@
+export interface CarData {
+    id: string;
+    brand: string;
+    name: string;
+    year: string;
+    km: string;
+    price: number;
+    specifications: string;
+    image: string;
+    color: string;
+    fuel: string;
+    fuelUrban: string;
+    fuelRoad: string;
+    dataSheet: string;
+}
